[refactor] Move serializers to  package
 - Use a common JsonConfig to avoid code duplication for events
